+++ src/window.rs
@@ -1,26 +1,26 @@
 ///  Anti-replay window, as described in https://tools.ietf.org/html/rfc6479
 /// TODO: use bitmap or generic-array?
-const NUM_BLOCKS: usize = 20;
-const BLOCK_LEN: usize = 5;
-const BITMAP_SIZE: usize = 1024; // 2^10, bits
-const BLOCK_SIZE: usize = 32; // 2^5, bits
-const WINDOW_SIZE: usize = BITMAP_SIZE - BLOCK_SIZE; // 2^10-2^5, bits
-const MAX_COUNTER: usize = usize::max_value() - WINDOW_SIZE - 1;
+const NUM_BLOCKS: u32 = 20;
+const BLOCK_LEN: u32 = 5;
+const BITMAP_SIZE: u32 = 1024; // 2^10, bits
+const BLOCK_SIZE: u32 = 32; // 2^5, bits
+const WINDOW_SIZE: u32 = BITMAP_SIZE - BLOCK_SIZE; // 2^10-2^5, bits
+const MAX_COUNTER: u32 = u32::max_value() - WINDOW_SIZE as u32 - 1;
 
 pub struct Window {
-    seen: [usize; NUM_BLOCKS],
-    counter: usize,
+    seen: [u32; NUM_BLOCKS as usize],
+    counter: u32,
 }
 
 impl Window {
     pub fn new() -> Self {
         Window {
-            seen: [0usize; NUM_BLOCKS],
-            counter: 0usize,
+            seen: [0u32; NUM_BLOCKS as usize],
+            counter: 0u32,
         }
     }
 
-    pub fn check_counter(&mut self, counter: usize) -> bool {
+    pub fn check_counter(&mut self, counter: u32) -> bool {
         if self.seen.len() == 0 {
             return true;
         }
@@ -35,29 +35,29 @@ impl Window {
             return false;
         }
 
-        let bit_loc_mask: usize = (1 << BLOCK_LEN as usize) - 1;
-        let bitmap_mask: usize = NUM_BLOCKS - 1;
-        let block_i: usize = counter >> BLOCK_LEN;
+        let bit_loc_mask: u32 = (1 << BLOCK_LEN as u32) - 1;
+        let bitmap_mask: u32 = NUM_BLOCKS - 1;
+        let block_i: u32 = counter >> BLOCK_LEN;
 
         // if nonce is not too old and > max seen nonce, update window, true
         if counter > self.counter {
-            let current_block_i: usize = self.counter >> BLOCK_LEN;
-            let diff: usize = usize::min(block_i - current_block_i, NUM_BLOCKS);
+            let current_block_i: u32 = self.counter >> BLOCK_LEN;
+            let diff: u32 = u32::min(block_i - current_block_i, NUM_BLOCKS);
             for i in 1..diff {
-                self.seen[(current_block_i + i) & bitmap_mask] = 0;
+                self.seen[((current_block_i + i) & bitmap_mask) as usize] = 0;
             }
             self.counter = counter;
         }
 
-        let actual_block_i: usize = block_i & bitmap_mask;
-        let bit_i: usize = counter & bit_loc_mask;
-        let bit_i_mask: usize = 1 << bit_i as usize;
+        let actual_block_i: u32 = block_i & bitmap_mask;
+        let bit_i: u32 = counter & bit_loc_mask;
+        let bit_i_mask: u32 = 1 << bit_i;
         // if counter is in window range and was seen before, false
-        if self.seen[actual_block_i] & bit_i_mask != 0 {
+        if self.seen[actual_block_i as usize] & bit_i_mask != 0 {
             return false;
         }
 
-        self.seen[actual_block_i] |= bit_i_mask;
+        self.seen[actual_block_i as usize] |= bit_i_mask;
         true
     }
 }
@@ -69,14 +69,14 @@ mod tests {
     #[test]
     fn test_check_counter() {
         let mut w = Window::new();
-        assert_eq!(false, w.check_counter(0usize), "msg 0");
-        assert_eq!(true, w.check_counter(1usize), "msg 1");
-        assert_eq!(false, w.check_counter(1usize), "msg 2");
-        assert_eq!(true, w.check_counter(5usize), "msg 3");
-        assert_eq!(true, w.check_counter(4usize), "msg 4");
+        assert_eq!(false, w.check_counter(0u32), "msg 0");
+        assert_eq!(true, w.check_counter(1u32), "msg 1");
+        assert_eq!(false, w.check_counter(1u32), "msg 2");
+        assert_eq!(true, w.check_counter(5u32), "msg 3");
+        assert_eq!(true, w.check_counter(4u32), "msg 4");
         assert_eq!(true, w.check_counter(BLOCK_LEN + 1), "msg 5");
         assert_eq!(true, w.check_counter(BLOCK_LEN * 2), "msg 6");
-        assert_eq!(false, w.check_counter(4usize), "msg 7");
+        assert_eq!(false, w.check_counter(4u32), "msg 7");
         assert_eq!(true, w.check_counter(BLOCK_LEN * 3), "msg 8");
         assert_eq!(true, w.check_counter(BLOCK_LEN * 2 - 1), "msg 9");
         assert_eq!(true, w.check_counter(BLOCK_LEN * 2 - 2), "msg 10");
@@ -85,10 +85,10 @@ mod tests {
         assert_eq!(true, w.check_counter(WINDOW_SIZE - 1), "msg 13");
         assert_eq!(false, w.check_counter(WINDOW_SIZE - 1), "msg 14");
         assert_eq!(true, w.check_counter(WINDOW_SIZE - 2), "msg 15");
-        assert_eq!(true, w.check_counter(2usize), "msg 16");
-        assert_eq!(false, w.check_counter(2usize), "msg 17");
+        assert_eq!(true, w.check_counter(2u32), "msg 16");
+        assert_eq!(false, w.check_counter(2u32), "msg 17");
         assert_eq!(true, w.check_counter(WINDOW_SIZE + 16), "msg 18");
-        assert_eq!(false, w.check_counter(3usize), "msg 19");
+        assert_eq!(false, w.check_counter(3u32), "msg 19");
         assert_eq!(false, w.check_counter(WINDOW_SIZE + 16), "msg 20");
         assert_eq!(true, w.check_counter(WINDOW_SIZE * 4), "msg 21");
         assert_eq!(
@@ -96,7 +96,7 @@ mod tests {
             w.check_counter(WINDOW_SIZE * 4 - (WINDOW_SIZE - 1)),
             "msg 22"
         );
-        assert_eq!(false, w.check_counter(10usize), "msg 23");
+        assert_eq!(false, w.check_counter(10u32), "msg 23");
         assert_eq!(
             true,
             w.check_counter(WINDOW_SIZE * 4 - WINDOW_SIZE),
@@ -126,6 +126,6 @@ mod tests {
         assert_eq!(false, w.check_counter(MAX_COUNTER + 2), "msg 34");
         assert_eq!(false, w.check_counter(MAX_COUNTER - 2), "msg 35");
         assert_eq!(true, w.check_counter(MAX_COUNTER - 3), "msg 36");
-        assert_eq!(false, w.check_counter(12usize), "msg 37");
+        assert_eq!(false, w.check_counter(12u32), "msg 37");
     }
 }
